+++ CHANGES.rst
@@ -1,7 +1,17 @@
 Changes
 *******
 
-0.1.0 (2018-07-04)
-==================
+0.2.0 (2018-11-29) Washington
+=============================
 
-* First release.
+* Provides generic RAVEN framework configuration
+* Process for GR4J-Cemaneige emulator
+* Process for HMETS emulator
+
+0.3.0 (2018-12-??)
+=============================
+
+* Adds process for MOHYSE emulator
+* Adds process for HBV-EC emulator
+
+